update: fixed issue with decoder producing a large diff.
The LZW function was starting from the beginning of the buffer
and processing data upto the end of the current chunk. This
produced very incorrect compressed data, which the decoder
silently decoded. This is now fixed.

An important thing to note here is that, we are using a vector
in the dedup function to maintain the mapping between sha fingerprint
and the chunk id, which gives us an O(n) access time, which is realllly
slow!! The unordered_map implementation according to us leads to
collisions, and hence it will require us to store the individual
chunks to make sure that even when a lookup succeeds the actual
chunks are the same.